Test and api update

+++ src/main/java/com/example/wirelesshome/model/ScheduleRequest.java
@@ -12,5 +12,5 @@ public class ScheduleRequest {
     private String name;
 
     private String type;
-    private int timer;
+    private long timer;
 }
